smart-todo-backend/requirements.txt
# Django Core Framework
Django==5.1.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2

# Database
psycopg2-binary==2.9.9

# AI Integration - Assignment Requirement
google-generativeai==0.8.2
openai==1.37.1

# Security & Authentication
django-allauth==0.63.6
PyJWT==2.8.0
cryptography==41.0.7

# HTTP & API
requests==2.32.3
urllib3==2.2.2
httpx==0.27.0

# Data Processing
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0

# Environment Configuration
python-dotenv==1.0.1
python-decouple==3.8

# Image & File Handling
Pillow==10.4.0

# Async & Background Tasks
celery==5.3.4
redis==5.0.8
channels==4.1.0

# Text Processing
nltk==3.8.1
textblob==0.17.1

# Content Processing
markdown==3.6
bleach==6.1.0

# Time & Scheduling
python-crontab==3.0.0
schedule==1.2.2

# Testing
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==26.0.0

# Monitoring & Debugging
sentry-sdk==2.12.0
django-extensions==3.2.3
